# backend/openapi.json
{"openapi":"3.1.0","info":{"title":"SAP PP/QM Job Collector API","version":"0.1.0"},"paths":{"/jobs":{"get":{"summary":"List Jobs","description":"Backward compatible: source, location, search, sort, limit, offset,\npage, page_size all behave exactly as in V15 if the new params\n(location_bucket, category, company, salary, posted_within) are\nomitted. Every new param is purely additive.\n\nRelevance scoring note: score_job() (app/relevance.py, UNCHANGED)\nis called here fresh per request, purely to support sort=relevance\nand the relevance_score field in each result. The score is never\nwritten back to the database.","operationId":"list_jobs_jobs_get","parameters":[{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by source, e.g. 'jooble'","title":"Source"},"description":"Filter by source, e.g. 'jooble'"},{"name":"location","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by location substring (legacy, exact substring match)","title":"Location"},"description":"Filter by location substring (legacy, exact substring match)"},{"name":"location_bucket","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by normalized location bucket: UAE|India|Germany|UK|Remote|Other","title":"Location Bucket"},"description":"Filter by normalized location bucket: UAE|India|Germany|UK|Remote|Other"},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by job category, e.g. 'SAP PP'. See /jobs/facets for the full list.","title":"Category"},"description":"Filter by job category, e.g. 'SAP PP'. See /jobs/facets for the full list."},{"name":"company","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Filter by exact company name","title":"Company"},"description":"Filter by exact company name"},{"name":"salary","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"any (default) | available | not_listed","title":"Salary"},"description":"any (default) | available | not_listed"},{"name":"posted_within","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"today | 7 | 30 | 90 (days)","title":"Posted Within"},"description":"today | 7 | 30 | 90 (days)"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"Search across title, company, location, and description","title":"Search"},"description":"Search across title, company, location, and description"},{"name":"sort","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"newest (default) | oldest | relevance | company | location","default":"newest","title":"Sort"},"description":"newest (default) | oldest | relevance | company | location"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":500,"default":100,"title":"Limit"}},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Offset"}},{"name":"page","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","minimum":1},{"type":"null"}],"description":"1-indexed page number; overrides limit/offset if given with page_size","title":"Page"},"description":"1-indexed page number; overrides limit/offset if given with page_size"},{"name":"page_size","in":"query","required":false,"schema":{"anyOf":[{"type":"integer","maximum":500,"minimum":1},{"type":"null"}],"description":"Results per page; overrides limit/offset if given with page","title":"Page Size"},"description":"Results per page; overrides limit/offset if given with page"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/facets":{"get":{"summary":"Jobs Facets","description":"New endpoint (additive -- does not change GET /jobs's existing\nbehavior or fields). Returns everything the sidebar and dashboard\nstats need in one call, computed from the same JOB_MAX_AGE_DAYS\nfloor as /jobs:\n\n    {\n      \"locations\": [{\"key\": \"All\", \"count\": N}, {\"key\": \"UAE\", \"count\": N}, ...],\n      \"categories\": [{\"key\": \"All Jobs\", \"count\": N}, {\"key\": \"SAP PP\", \"count\": N}, ...],\n      \"companies\": [{\"name\": \"Accenture\", \"count\": N}, ...],   # top 100 by count, desc\n      \"stats\": {\"total\": N, \"new_today\": N, \"remote\": N, \"uae\": N,\n                 \"india\": N, \"germany\": N, \"hr\": N, \"sap\": N}\n    }\n\nProper faceted-search semantics: each facet's counts are computed\nwith every OTHER active filter applied, but not its own dimension\n-- so switching the location filter shows how many jobs are in\neach location bucket given the current category/company/salary/\nposted_within/search selections, without the current location\nselection itself constraining those counts. `stats`, by contrast,\nreflects ALL currently active filters together (it's a summary of\n\"what you're looking at right now\", not a facet picker).","operationId":"jobs_facets_jobs_facets_get","parameters":[{"name":"source","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source"}},{"name":"location_bucket","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Location Bucket"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"company","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company"}},{"name":"salary","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Salary"}},{"name":"posted_within","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Posted Within"}},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Search"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profiles":{"post":{"summary":"Create Profile","operationId":"create_profile_profiles_post","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string","title":"Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"summary":"List Profiles","operationId":"list_profiles_profiles_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/profiles/{profile_id}/preferences":{"put":{"summary":"Update Preferences","operationId":"update_preferences_profiles__profile_id__preferences_put","parameters":[{"name":"profile_id","in":"path","required":true,"schema":{"type":"integer","title":"Profile Id"}},{"name":"preferred_job_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Preferred Job Type"}},{"name":"expected_salary","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Expected Salary"}},{"name":"notice_period","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Notice Period"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_update_preferences_profiles__profile_id__preferences_put"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profiles/{profile_id}/resume":{"post":{"summary":"Upload Resume","operationId":"upload_resume_profiles__profile_id__resume_post","parameters":[{"name":"profile_id","in":"path","required":true,"schema":{"type":"integer","title":"Profile Id"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_resume_profiles__profile_id__resume_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profiles/{profile_id}/resume/{resume_id}/review":{"post":{"summary":"Review Resume","description":"decision: yes | edit | reject","operationId":"review_resume_profiles__profile_id__resume__resume_id__review_post","parameters":[{"name":"profile_id","in":"path","required":true,"schema":{"type":"integer","title":"Profile Id"}},{"name":"resume_id","in":"path","required":true,"schema":{"type":"integer","title":"Resume Id"}},{"name":"decision","in":"query","required":true,"schema":{"type":"string","title":"Decision"}}],"requestBody":{"content":{"application/json":{"schema":{"anyOf":[{"type":"object","additionalProperties":true},{"type":"null"}],"title":"Edits"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/profiles/{profile_id}/resume/active":{"get":{"summary":"Get Active Resume","operationId":"get_active_resume_profiles__profile_id__resume_active_get","parameters":[{"name":"profile_id","in":"path","required":true,"schema":{"type":"integer","title":"Profile Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/jobs/{job_id}/match":{"get":{"summary":"Job Match","operationId":"job_match_jobs__job_id__match_get","parameters":[{"name":"job_id","in":"path","required":true,"schema":{"type":"integer","title":"Job Id"}},{"name":"profile_id","in":"query","required":true,"schema":{"type":"integer","title":"Profile Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/dashboard/{profile_id}":{"get":{"summary":"Profile Dashboard","operationId":"profile_dashboard_dashboard__profile_id__get","parameters":[{"name":"profile_id","in":"path","required":true,"schema":{"type":"integer","title":"Profile Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/collector-status":{"get":{"summary":"Collector Status","operationId":"collector_status_collector_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/export":{"post":{"summary":"Export Jobs","operationId":"export_jobs_export_post","parameters":[{"name":"format","in":"query","required":false,"schema":{"type":"string","default":"csv","title":"Format"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"array","items":{"type":"string"},"title":"Job Urls"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_update_preferences_profiles__profile_id__preferences_put":{"properties":{"preferred_countries":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Countries"},"preferred_cities":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Cities"},"preferred_modules":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Preferred Modules"}},"type":"object","title":"Body_update_preferences_profiles__profile_id__preferences_put"},"Body_upload_resume_profiles__profile_id__resume_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_resume_profiles__profile_id__resume_post"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}